+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+## [1.3.0] - 2025-07-15
 ### Removed
 - `DataDescriber.{output_path,csv_dir,yaml_dir}` paths. Now specific paths of
   where to save data is provided to `save*` methods.
@@ -16,7 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `DataDescriber` `save*` methods expect specific paths of where data is saved.
 
 ### Added
-- Add `DataDescriptor` JSON serialization.
+- `DataDescriptor` JSON serialization.
+- Serialized files can be input files to the CLI.
 
 
 ## [1.2.3] - 2025-07-15
@@ -187,7 +189,8 @@ Downstream moderate risk update release.
 
 
 <!-- links -->
-[Unreleased]: https://github.com/plandes/datdesc/compare/v1.2.3...HEAD
+[Unreleased]: https://github.com/plandes/datdesc/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/plandes/datdesc/compare/v1.2.3...v1.3.0
 [1.2.3]: https://github.com/plandes/datdesc/compare/v1.2.2...v1.2.3
 [1.2.2]: https://github.com/plandes/datdesc/compare/v1.2.1...v1.2.2
 [1.2.1]: https://github.com/plandes/datdesc/compare/v1.2.0...v1.2.1